requirements.txt
Flask
numpy
opencv-python
Pillow
scikit-image
scipy
tensorflow
